libavformat/rtsp: Make source specific multicast work for rtsp streams
by first changing the RTSPSource to track the destination address
obtained from the source filter. For each RTSPStream, only add the source
filter from the sdp if sdp_ip string matches source-filter's destination
address.

Before issuing the setup request, change the lower_transport to
multicast if the sdp_ip is a multicast address.

Change the multicast case to append sources (from the source-filter) to
the rtp url to make the source specific multicast join work.

Support wildcard source-filter destinations.
Change source-filter processing to only happen after the first
media description was processed and either when the next one starts
or when we have reached the end of the sdp.

This handles a wider variety of cases and doesn't rely on any optional
media descriptions.